升级 MybatisPlus 至 3.4.3.3

+++ demo-SpringBoot-MybatisPlus/src/test/java/com/maxqiu/demo/base/TestUser.java
@@ -14,7 +14,7 @@ import com.maxqiu.demo.entity.User;
 
 /**
  * 测试 Entity 层 / 测试 AR 模式
- * 
+ *
  * @author Max_Qiu
  */
 @SpringBootTest
@@ -164,7 +164,7 @@ public class TestUser {
         // SELECT COUNT( * ) FROM smp_user WHERE (age > ? AND age < ?)
         LambdaQueryWrapper<User> queryWrapper = Wrappers.lambdaQuery();
         queryWrapper.gt(User::getAge, 0).lt(User::getAge, 100);
-        int count = new User().selectCount(queryWrapper);
+        long count = new User().selectCount(queryWrapper);
         System.out.println(count);
     }
 
@@ -191,5 +191,4 @@ public class TestUser {
         List<User> users = userPage.getRecords();
         users.forEach(System.out::println);
     }
-
 }
